Refactor health check resources to properly subclass abstract resource.
